+++ .gitignore
@@ -48,6 +48,7 @@ htmlcov/
 # Django stuff:
 *.log
 local_settings.py
+db.sqlite3
 
 # Flask stuff:
 instance/
