+++ api/tests.py
@@ -1,3 +1,6 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
 from django.test import TestCase
 
 # Create your tests here.
