--- api_v1_0/apps.py
@@ -1,7 +1,0 @@
-from __future__ import unicode_literals
-
-from django.apps import AppConfig
-
-
-class ApiV10Config(AppConfig):
-    name = 'api_v1_0'
