+++ application/settings.py
@@ -28,18 +28,18 @@ sys.path.insert(0, functions_dir)
 
 # ==================
 
-def assign_env_value(var_name):
-	if var_name in os.environ:
-		return getenv.env(var_name)
-	else:
-		sys.exit(var_name + " is not defined in the environment variables")
+def getEnvVariable(var_name):
+  if var_name in os.environ:
+    return getenv.env(var_name)
+  else:
+    sys.exit(var_name + " is not defined in the environment variables")
 
 # Quick-start development settings - unsuitable for production
 # See https://docs.djangoproject.com/en/1.10/howto/deployment/checklist/
 
 
-SECRET_KEY = assign_env_value('SECRET_KEY')
-DEBUG = assign_env_value('DEBUG')
+SECRET_KEY = getEnvVariable('SECRET_KEY')
+DEBUG = getEnvVariable('DEBUG')
 
 
 # Application definition
@@ -51,8 +51,8 @@ INSTALLED_APPS = (
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-	'api_v1_0',
-	'rest_framework',
+    'api',
+    'rest_framework',
     'oauth2_provider',
 )
 
@@ -92,22 +92,9 @@ WSGI_APPLICATION = 'application.wsgi.application'
 # Database
 # https://docs.djangoproject.com/en/1.9/ref/settings/#databases
 
-if "test" in sys.argv:
-	DATABASES = {
-		'default': {
-			'ENGINE': 'django.db.backends.sqlite3',
-			'NAME': os.path.join(BASE_DIR, 'db.sqlite3'),
-			'CONN_MAX_AGE': 500,
-			'TEST' :
-			{
-				'NAME': os.path.join(BASE_DIR, 'db_test.sqlite3'),
-			}
-		},
-	}
-else:
-	DATABASES = {
-		'default': dj_database_url.config(default=assign_env_value('DATABASE_URL'), conn_max_age=500),
-	}
+DATABASES = {
+'default': dj_database_url.config(default=getEnvVariable('DATABASE_URL'), conn_max_age=500),
+}
 
 # Password validation
 # https://docs.djangoproject.com/en/1.9/ref/settings/#auth-password-validators
@@ -170,3 +157,9 @@ REST_FRAMEWORK = {
         'rest_framework.permissions.IsAuthenticated',
     )
 }
+
+try:
+    from local_settings import *
+except Exception as e:
+    print e
+    pass
