+++ application/urls.py
@@ -1,3 +1,6 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
 from django.conf.urls import include, url
 from django.contrib.auth.models import User, Group
 from django.contrib import admin
