+++ functions/time_funcs.py
@@ -4,7 +4,6 @@
 from __future__ import unicode_literals
 
 import datetime, time
-# Create your views here.
 
 def get_timestamp():
-	return datetime.datetime.fromtimestamp(time.time()).strftime('%Y-%m-%d %H:%M:%S')
+    return datetime.datetime.fromtimestamp(time.time()).strftime('%Y-%m-%d %H:%M:%S')
